+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/DefaultNativeConnectionProvider.java
@@ -21,22 +21,22 @@ import com.ericsson.bss.cassandra.ecchronos.connection.impl.LocalNativeConnectio
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Properties;
-
 public class DefaultNativeConnectionProvider implements NativeConnectionProvider
 {
     private static final Logger LOG = LoggerFactory.getLogger(DefaultNativeConnectionProvider.class);
 
     private final LocalNativeConnectionProvider myLocalNativeConnectionProvider;
 
-    public DefaultNativeConnectionProvider(Properties properties) throws ConfigurationException
+    public DefaultNativeConnectionProvider(Config config)
     {
-        DefaultNativeConnectionProperties connectionProperties = DefaultNativeConnectionProperties.from(properties);
-        LOG.info("Connecting through CQL using {}", connectionProperties);
+        Config.NativeConnection nativeConfig = config.getConnectionConfig().getCql();
+        String host = nativeConfig.getHost();
+        int port = nativeConfig.getPort();
+        LOG.info("Connecting through CQL using {}:{}", host, port);
 
         myLocalNativeConnectionProvider = LocalNativeConnectionProvider.builder()
-                .withLocalhost(connectionProperties.getNativeHost())
-                .withPort(connectionProperties.getNativePort())
+                .withLocalhost(host)
+                .withPort(port)
                 .build();
     }