+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/ECChronosInternals.java
@@ -19,13 +19,7 @@ import com.datastax.driver.core.Metadata;
 import com.ericsson.bss.cassandra.ecchronos.connection.JmxConnectionProvider;
 import com.ericsson.bss.cassandra.ecchronos.connection.NativeConnectionProvider;
 import com.ericsson.bss.cassandra.ecchronos.connection.StatementDecorator;
-import com.ericsson.bss.cassandra.ecchronos.core.CASLockFactory;
-import com.ericsson.bss.cassandra.ecchronos.core.HostStates;
-import com.ericsson.bss.cassandra.ecchronos.core.HostStatesImpl;
-import com.ericsson.bss.cassandra.ecchronos.core.JmxProxyFactory;
-import com.ericsson.bss.cassandra.ecchronos.core.JmxProxyFactoryImpl;
-import com.ericsson.bss.cassandra.ecchronos.core.TableStorageStates;
-import com.ericsson.bss.cassandra.ecchronos.core.TableStorageStatesImpl;
+import com.ericsson.bss.cassandra.ecchronos.core.*;
 import com.ericsson.bss.cassandra.ecchronos.core.metrics.TableRepairMetrics;
 import com.ericsson.bss.cassandra.ecchronos.core.metrics.TableRepairMetricsImpl;
 import com.ericsson.bss.cassandra.ecchronos.core.scheduling.RunPolicy;
@@ -38,7 +32,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.Closeable;
-import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
 public class ECChronosInternals implements Closeable
@@ -58,12 +51,9 @@ public class ECChronosInternals implements Closeable
 
     private final CASLockFactory myLockFactory;
 
-    public ECChronosInternals(Properties configuration, NativeConnectionProvider nativeConnectionProvider,
+    public ECChronosInternals(Config configuration, NativeConnectionProvider nativeConnectionProvider,
                               JmxConnectionProvider jmxConnectionProvider, StatementDecorator statementDecorator)
-            throws ConfigurationException
     {
-        StatisticsProperties statisticsProperties = StatisticsProperties.from(configuration);
-
         myJmxProxyFactory = JmxProxyFactoryImpl.builder()
                 .withJmxConnectionProvider(jmxConnectionProvider)
                 .build();
@@ -72,13 +62,11 @@ public class ECChronosInternals implements Closeable
                 .withJmxProxyFactory(myJmxProxyFactory)
                 .build();
 
-        CASLockFactoryProperties casLockFactoryProperties = CASLockFactoryProperties.from(configuration);
-
         myLockFactory = CASLockFactory.builder()
                 .withNativeConnectionProvider(nativeConnectionProvider)
                 .withHostStates(myHostStatesImpl)
                 .withStatementDecorator(statementDecorator)
-                .withKeyspaceName(casLockFactoryProperties.getKeyspaceName())
+                .withKeyspaceName(configuration.getLockFactory().getCas().getKeyspace())
                 .build();
 
         Host host = nativeConnectionProvider.getLocalHost();
@@ -86,7 +74,7 @@ public class ECChronosInternals implements Closeable
 
         myReplicatedTableProvider = new ReplicatedTableProviderImpl(host, metadata);
 
-        if (statisticsProperties.isEnabled())
+        if (configuration.getStatistics().isEnabled())
         {
             myTableStorageStatesImpl = TableStorageStatesImpl.builder()
                     .withReplicatedTableProvider(myReplicatedTableProvider)
@@ -95,7 +83,7 @@ public class ECChronosInternals implements Closeable
 
             myTableRepairMetricsImpl = TableRepairMetricsImpl.builder()
                     .withTableStorageStates(myTableStorageStatesImpl)
-                    .withStatisticsDirectory(statisticsProperties.getStatisticsDirectory().toString())
+                    .withStatisticsDirectory(configuration.getStatistics().getDirectory().toString())
                     .build();
         }
         else
@@ -104,11 +92,10 @@ public class ECChronosInternals implements Closeable
             myTableRepairMetricsImpl = null;
         }
 
-        SchedulerProperties schedulerProperties = SchedulerProperties.from(configuration);
-
         myScheduleManagerImpl = ScheduleManagerImpl.builder()
                 .withLockFactory(myLockFactory)
-                .withRunInterval(schedulerProperties.getRunInterval(), schedulerProperties.getTimeUnit())
+                .withRunInterval(configuration.getScheduler().getFrequency().getInterval(TimeUnit.MILLISECONDS),
+                        TimeUnit.MILLISECONDS)
                 .build();
     }